// src/shared/api/generated/model/agendaResponseDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Yoclick API
 * API multi-centro de reservas en marca blanca. Contrato canónico para la app.
 * OpenAPI spec version: 0.1.0
 */
import type { AgendaResponseDtoEntriesItem } from './agendaResponseDtoEntriesItem.ts';

export interface AgendaResponseDto {
  /** @pattern ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))$ */
  date: string;
  timezone: string;
  entries: AgendaResponseDtoEntriesItem[];
}