// src/shared/api/generated/model/agendaResponseDtoEntriesItem.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Yoclick API
 * API multi-centro de reservas en marca blanca. Contrato canónico para la app.
 * OpenAPI spec version: 0.1.0
 */
import type { AgendaResponseDtoEntriesItemBooking } from './agendaResponseDtoEntriesItemBooking.ts';
import type { AgendaResponseDtoEntriesItemClient } from './agendaResponseDtoEntriesItemClient.ts';

export type AgendaResponseDtoEntriesItem = {
  booking: AgendaResponseDtoEntriesItemBooking;
  client: AgendaResponseDtoEntriesItemClient;
};